requirements.txt
django==1.10.5
djangorestframework==3.6.3
django-rest-swagger==2.1.2
pillow
django-imagekit==4.0